Issue #1 Refactoring routes to own folder
